hello-prodify/requirements.txt
fastapi
uvicorn
google-auth
requests
google-cloud-firestore

